+++ src/components/modules/toolbar/blockSettings.ts
@@ -29,7 +29,7 @@ export default class BlockSettings extends Module {
    * Block Settings CSS
    * @return {{wrapper, wrapperOpened, toolSettings, defaultSettings, button}}
    */
-  private static get CSS() {
+  public get CSS() {
     return {
       // Settings Panel
       wrapper: 'ce-settings',
@@ -38,6 +38,9 @@ export default class BlockSettings extends Module {
       defaultSettings: 'ce-settings__default-zone',
 
       button: 'ce-settings__button',
+
+      focusedButton : 'ce-settings__button--focused',
+      focusedButtonAnimated: 'ce-settings__button--focused-animated',
     };
   }
 
@@ -46,7 +49,7 @@ export default class BlockSettings extends Module {
    * @returns {boolean}
    */
   public get opened(): boolean {
-    return this.nodes.wrapper.classList.contains(BlockSettings.CSS.wrapperOpened);
+    return this.nodes.wrapper.classList.contains(this.CSS.wrapperOpened);
   }
 
   /**
@@ -59,6 +62,16 @@ export default class BlockSettings extends Module {
   };
 
   /**
+   * List of buttons
+   */
+  private buttons: HTMLElement[] = [];
+
+  /**
+   * Index of active button
+   */
+  private focusedButtonIndex: number = -1;
+
+  /**
    * Panel with block settings with 2 sections:
    *  - Tool's Settings
    *  - Default Settings [Move, Remove, etc]
@@ -66,10 +79,10 @@ export default class BlockSettings extends Module {
    * @return {Element}
    */
   public make(): void {
-    this.nodes.wrapper = $.make('div', BlockSettings.CSS.wrapper);
+    this.nodes.wrapper = $.make('div', this.CSS.wrapper);
 
-    this.nodes.toolSettings = $.make('div', BlockSettings.CSS.toolSettings);
-    this.nodes.defaultSettings = $.make('div', BlockSettings.CSS.defaultSettings);
+    this.nodes.toolSettings = $.make('div', this.CSS.toolSettings);
+    this.nodes.defaultSettings = $.make('div', this.CSS.defaultSettings);
 
     $.append(this.nodes.wrapper, [this.nodes.toolSettings, this.nodes.defaultSettings]);
   }
@@ -78,7 +91,7 @@ export default class BlockSettings extends Module {
    * Open Block Settings pane
    */
   public open(): void {
-    this.nodes.wrapper.classList.add(BlockSettings.CSS.wrapperOpened);
+    this.nodes.wrapper.classList.add(this.CSS.wrapperOpened);
 
     /**
      * Fill Tool's settings
@@ -98,7 +111,7 @@ export default class BlockSettings extends Module {
    * Close Block Settings pane
    */
   public close(): void {
-    this.nodes.wrapper.classList.remove(BlockSettings.CSS.wrapperOpened);
+    this.nodes.wrapper.classList.remove(this.CSS.wrapperOpened);
 
     /** Clear settings */
     this.nodes.toolSettings.innerHTML = '';
@@ -106,8 +119,66 @@ export default class BlockSettings extends Module {
 
     /** Tell to subscribers that block settings is closed */
     this.Editor.Events.emit(this.events.closed);
+
+    /** Clear cached buttons */
+    this.buttons = [];
+
+    /** Clear focus on active button */
+    this.focusedButtonIndex = -1;
+
   }
 
+  /**
+   * Returns Tools Settings and Default Settings
+   * @return {HTMLElement[]}
+   */
+  public get blockTunesButtons(): HTMLElement[] {
+    /**
+     * Return from cache
+     * if exists
+     */
+    if (this.buttons.length !== 0) {
+      return this.buttons;
+    }
+
+    const toolSettings = this.nodes.toolSettings.querySelectorAll(`.${this.Editor.StylesAPI.classes.settingsButton}`);
+    const defaultSettings = this.nodes.defaultSettings.querySelectorAll(`.${this.CSS.button}`);
+
+    toolSettings.forEach((item, index) => {
+      this.buttons.push((item as HTMLElement));
+      if (item.classList.contains(this.CSS.focusedButton)) {
+        this.focusedButtonIndex = index;
+      }
+    });
+
+    defaultSettings.forEach((item) => {
+      this.buttons.push((item as HTMLElement));
+    });
+
+    return this.buttons;
+  }
+
+  /**
+   * Leaf Block Tunes
+   * @param {string} direction
+   */
+  public leaf(direction: string = 'right'): void {
+    this.focusedButtonIndex = $.leafNodesAndReturnIndex(
+      this.blockTunesButtons, this.focusedButtonIndex, direction, this.CSS.focusedButton,
+    );
+  }
+
+  /**
+   * Returns active button HTML element
+   * @return {HTMLElement}
+   */
+  public get focusedButton(): HTMLElement {
+    if (this.focusedButtonIndex === -1) {
+      return null;
+    }
+
+    return (this.buttons[this.focusedButtonIndex] as HTMLElement);
+  }
   /**
    * Add Tool's settings
    */